Added deleteHandler function for Documents
Added a delete property to the Documents with a default of false.

Added deleteHandler to set the delete property of Document model to
true.

I'm planning on implementing a scheduled task that, once a day, looks for all
documents tagged for deletion and then does a batch deletion on
cloudinary

Created tests for response status code